requirements.txt
click
tqdm
requests
Jinja2
prompt_toolkit
ruamel.yaml